Refactor: Move settings into respective view file
